feat: 支持自定义配置(#master)

+++ src/upload-dist.js
@@ -1,35 +1,23 @@
-const path = require('path')
-const fs = require('fs')
+import * as path from 'path'
+import * as fs from 'fs'
+import * as dotenv from 'dotenv'
+import SFTPClient from 'ssh2-sftp-client'
+
 const resolvePath = (dir) => {
   return path.join(process.cwd(), dir)
 }
-// https://www.npmjs.com/package/dotenv
-const dotenv = require('dotenv').config({
-  debug: process.env.DEBUG,
-  path: resolvePath('.env.local')
-})
 
 // https://www.npmjs.com/package/ssh2-sftp-client#sec-1
-let Client = require('ssh2-sftp-client')
-let sftp = new Client()
+// let SFTPClient = require('ssh2-sftp-client')
+let sftp = new SFTPClient()
 
-const config = {
-  host: process.env.SFTP_SERVER,
-  username: process.env.SFTP_USER,
-  password: process.env.SFTP_PASSWORD,
-  port: process.env.SFTP_PORT || 22
-}
-
-let srcDir = resolvePath(process.env.BUILD_DIR)
-let dstDir = process.env.DESTINE_DIR
-
-const isDirectory = str => {
+const isDirectory = (str) => {
   return str === 'd'
 }
 
-async function isExistDist() {
+async function isExistDist(srcDir) {
   return new Promise((resolve, reject) => {
-    fs.access(srcDir, fs.constants.F_OK, err => {
+    fs.access(srcDir, fs.constants.F_OK, (err) => {
       if (err) {
         console.error('本地打包文件不存在', err)
         process.exit(1)
@@ -39,17 +27,57 @@ async function isExistDist() {
   })
 }
 
-export function run() {
-  isExistDist().then(() => {
+function initConfig(configFile = '.env.local') {
+  // https://www.npmjs.com/package/dotenv
+  dotenv.config({
+    debug: process.env.DEBUG,
+    path: resolvePath(configFile),
+  })
+  const config = {
+    host: process.env.SFTP_SERVER,
+    username: process.env.SFTP_USER,
+    password: process.env.SFTP_PASSWORD,
+    port: process.env.SFTP_PORT || 22,
+    ignore: '(.git|node_modules)',
+    srcDir: process.env.BUILD_DIR,
+    dstDir: process.env.DESTINE_DIR,
+  }
+  return config
+}
+
+export function run(options) {
+  let config = initConfig(options.file)
+  let { srcDir, dstDir } = config
+
+  Object.keys(options).forEach((key) => {
+    if (options[key]) {
+      config[key] = options[key]
+    }
+  })
+  if (config.dstDir) {
+    dstDir = config.dstDir
+  }
+  if (config.srcDir) {
+    srcDir = resolvePath(config.srcDir)
+  }
+
+  isExistDist(srcDir).then(() => {
     sftp
       .connect(config)
       .then(() => {
         return sftp.list(dstDir)
       })
-      .then(files => {
+      .then((files) => {
+        if (config.isTest) {
+          files.forEach((item) => {
+            console.log(`Dirty delete file: ${item.longname}`)
+          })
+          console.log('Testing connect successed.')
+          return
+        }
         // 获得远程部署目录文件信息
         // 遍历删除目录下的文件
-        let promise = files.map(file => {
+        let promise = files.map((file) => {
           let filepath = path.join(dstDir, file.name).replace(/\\/gi, '/')
           if (isDirectory(file.type)) {
             return sftp.rmdir(filepath, true)
@@ -59,19 +87,26 @@ export function run() {
         })
         return Promise.all(promise)
       })
-      .then(data => {
-        sftp.on('upload', info => {
+      .then((data) => {
+        sftp.on('upload', (info) => {
           console.log(`Listener: Uploaded ${info.source}`)
         })
       })
       .then(() => {
-        return sftp.uploadDir(srcDir, dstDir)
+        if (config.isTest) {
+          return srcDir + '  ' + dstDir
+        }
+        return sftp.uploadDir(srcDir, dstDir, {
+          filter: (item) => {
+            return !new RegExp(config.ignore, 'g').test(item)
+          },
+        })
       })
-      .then(data => {
+      .then((data) => {
         console.log('上传完成', data)
       })
-      .catch(err => {
-        console.log(err, 'catch error')
+      .catch((err) => {
+        console.log(err)
       })
       .finally(() => {
         // 断开连接
